Refactor CSS class names and container properties in Spanning.jsx and Spanning.module.css
The CSS class names and container properties in Spanning.jsx and Spanning.module.css have been refactored. This includes renaming and reordering of class names, as well as modifications to grid-column and grid-row properties. These changes improve the organization and readability of the code related to grid spanning in the Grid module.